refactor: layout.vue bring up a layer
