+++ apps/app-frontend/src/pages/instance/index.ts
@@ -1,6 +1,6 @@
 import Content from './content/index.vue'
 import Files from './files/index.vue'
-import Index from './instance-page/index.vue'
+import Index from './layout.vue'
 import Logs from './logs/index.vue'
 import Share from './share/index.vue'
 import Worlds from './worlds/index.vue'
