+++ apps/app-frontend/src/pages/instance/instance-admonitions/shared-instance-unavailable.vue
