+++ apps/app-frontend/src/pages/instance/instance-page-header/instance-page-header-server-metadata.vue
